% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_edit.R
\name{callEditsFromConsensus}
\alias{callEditsFromConsensus}
\title{Call edits from aligned consensus sequences}
\usage{
callEditsFromConsensus(align_df, ref_targets, confidence_df = NULL, out_dir)
}
\arguments{
\item{align_df}{Output from alignConsensusToRef.}

\item{ref_targets}{Output from prepRefTargets.}

\item{confidence_df}{Optional confidence metrics from evalConfidence.}

\item{out_dir}{Output directory.}
}
\value{
Data frame of edit calls per sample/cluster/target.
}
\description{
Call edits from aligned consensus sequences
}
