% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_demultiplex.R
\name{.align_end_score}
\alias{.align_end_score}
\title{Align a short query to a read-end window and return scores}
\usage{
.align_end_score(query, subject)
}
\description{
Align a short query to a read-end window and return scores
}
\keyword{internal}
