% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_edit.R
\name{.build_refless_editcall_summary}
\alias{.build_refless_editcall_summary}
\title{Build editViewer-compatible summary table from refless edit-calling results}
\usage{
.build_refless_editcall_summary(per_sample)
}
\description{
Build editViewer-compatible summary table from refless edit-calling results
}
\keyword{internal}
