% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_demultiplex.R
\name{.classify_demultiplex}
\alias{.classify_demultiplex}
\title{Classify demultiplex assignment from per-read scores}
\usage{
.classify_demultiplex(
  scores,
  t_high = 0.85,
  t_low = 0.7,
  delta = 0.05,
  min_end_id_high = 0.8
)
}
\description{
Classify demultiplex assignment from per-read scores
}
\keyword{internal}
