% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_edit.R
\name{.classify_edit_window}
\alias{.classify_edit_window}
\title{Classify edit pattern from consensus vs intact window}
\usage{
.classify_edit_window(consensus_window, ref_window, intact_ref)
}
\description{
Classify edit pattern from consensus vs intact window
}
\keyword{internal}
