% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_index.R
\name{.edit_distance}
\alias{.edit_distance}
\title{Pairwise edit distance between DNA strings (unit-cost)}
\usage{
.edit_distance(a, b)
}
\description{
Pairwise edit distance between DNA strings (unit-cost)
}
\keyword{internal}
