% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_edit.R
\name{.extract_aligned_window}
\alias{.extract_aligned_window}
\title{Extract edit-site window sequence from a global alignment}
\usage{
.extract_aligned_window(aln, window_start, window_end)
}
\description{
Extract edit-site window sequence from a global alignment
}
\keyword{internal}
