% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_report.R
\name{.flag_contaminant_clusters}
\alias{.flag_contaminant_clusters}
\title{Identify likely contaminant clusters (small fraction, low support)}
\usage{
.flag_contaminant_clusters(consensus_df, confidence_df, min_fraction = 0.05)
}
\description{
Identify likely contaminant clusters (small fraction, low support)
}
\keyword{internal}
