% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_edit.R
\name{.map_short_seqs_minimap}
\alias{.map_short_seqs_minimap}
\title{Map short sequences to reference with minimap2 and parse PAF hits}
\usage{
.map_short_seqs_minimap(query_fa, ref_fa, out_dir, sif_path = NULL)
}
\description{
Map short sequences to reference with minimap2 and parse PAF hits
}
\keyword{internal}
