% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_chimera.R
\name{.parse_derep_uc}
\alias{.parse_derep_uc}
\title{Parse derep UC file: map each read to its derep representative}
\usage{
.parse_derep_uc(uc_fn)
}
\description{
Parse derep UC file: map each read to its derep representative
}
\keyword{internal}
