% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_cluster.R
\name{.parse_mmseqs_cluster_tsv}
\alias{.parse_mmseqs_cluster_tsv}
\title{Parse MMseqs2 easy-linclust cluster TSV}
\usage{
.parse_mmseqs_cluster_tsv(tsv_fn)
}
\description{
Parse MMseqs2 easy-linclust cluster TSV
}
\keyword{internal}
