% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_cluster.R
\name{.parse_vsearch_uc}
\alias{.parse_vsearch_uc}
\title{Parse vsearch UC output into read->cluster mapping}
\usage{
.parse_vsearch_uc(uc_fn, read_ids)
}
\description{
Parse vsearch UC output into read->cluster mapping
}
\keyword{internal}
