% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_consensus.R
\name{.pileup_consensus}
\alias{.pileup_consensus}
\title{Majority consensus from aligned sequences (simple column-wise vote)}
\usage{
.pileup_consensus(seqs)
}
\description{
Majority consensus from aligned sequences (simple column-wise vote)
}
\keyword{internal}
