% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_consensus.R
\name{.pileup_support}
\alias{.pileup_support}
\title{Per-position support rates from global alignment to consensus}
\usage{
.pileup_support(seqs, consensus)
}
\description{
Per-position support rates from global alignment to consensus
}
\keyword{internal}
