% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_apptainer.R
\name{.read_primers}
\alias{.read_primers}
\title{Read primer sequences from CSV (col1=id, col2=seq)}
\usage{
.read_primers(primer_list)
}
\description{
Read primer sequences from CSV (col1=id, col2=seq)
}
\keyword{internal}
