% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_cluster.R
\name{.refine_clusters_minimap}
\alias{.refine_clusters_minimap}
\title{Refine clusters by re-grouping reads with high minimap2 identity to cluster medoid}
\usage{
.refine_clusters_minimap(cluster_df, fasta_fn, out_dir, sif_path = NULL)
}
\description{
Refine clusters by re-grouping reads with high minimap2 identity to cluster medoid
}
\keyword{internal}
