% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_apptainer.R
\name{.refless_container_runner}
\alias{.refless_container_runner}
\title{Detect apptainer or singularity runner}
\usage{
.refless_container_runner()
}
\description{
Detect apptainer or singularity runner
}
\keyword{internal}
