% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_edit.R
\name{.refless_genotype_to_eval}
\alias{.refless_genotype_to_eval}
\title{Classify refless genotype string into editViewer evaluation levels}
\usage{
.refless_genotype_to_eval(genotype)
}
\description{
Classify refless genotype string into editViewer evaluation levels
}
\keyword{internal}
