% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_apptainer.R
\name{.refless_read_seqs}
\alias{.refless_read_seqs}
\title{Read FASTA or FASTQ sequences (format from file extension)}
\usage{
.refless_read_seqs(path)
}
\description{
Read FASTA or FASTQ sequences (format from file extension)
}
\keyword{internal}
