% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_apptainer.R
\name{.refless_run_shell}
\alias{.refless_run_shell}
\title{Run shell command via Apptainer or host}
\usage{
.refless_run_shell(cmd, work_dir = getwd(), sif_path = NULL)
}
\description{
Run shell command via Apptainer or host
}
\keyword{internal}
