% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_apptainer.R
\name{.refless_run_umap_script}
\alias{.refless_run_umap_script}
\title{Run UMAP pre-cluster script on host Python or in container}
\usage{
.refless_run_umap_script(fasta_fn, out_tsv, min_samples = 50, sif_path = NULL)
}
\description{
Run UMAP pre-cluster script on host Python or in container
}
\keyword{internal}
