% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_apptainer.R
\name{.refless_tool_bin}
\alias{.refless_tool_bin}
\title{Resolve executable for an external tool}
\usage{
.refless_tool_bin(tool)
}
\description{
Resolve executable for an external tool
}
\keyword{internal}
