% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_apptainer.R
\name{.refless_tool_candidates}
\alias{.refless_tool_candidates}
\title{Known host locations for external tools}
\usage{
.refless_tool_candidates(tool)
}
\description{
Known host locations for external tools
}
\keyword{internal}
