% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_apptainer.R
\name{.refless_umap_script}
\alias{.refless_umap_script}
\title{Path to bundled UMAP pre-cluster script}
\usage{
.refless_umap_script()
}
\description{
Path to bundled UMAP pre-cluster script
}
\keyword{internal}
