% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_cluster.R
\name{.run_mmseqs_cluster}
\alias{.run_mmseqs_cluster}
\title{Run MMseqs2 easy-linclust via Apptainer or host}
\usage{
.run_mmseqs_cluster(
  fasta_fn,
  out_dir,
  min_seq_id = 0.85,
  c_cov = 0.8,
  sif_path = NULL
)
}
\description{
Run MMseqs2 easy-linclust via Apptainer or host
}
\keyword{internal}
