% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_chimera.R
\name{.run_uchime_denovo}
\alias{.run_uchime_denovo}
\title{Run vsearch derep + uchime_denovo}
\usage{
.run_uchime_denovo(fasta_fn, out_dir, abskew = 2, sif_path = NULL)
}
\description{
Run vsearch derep + uchime_denovo
}
\keyword{internal}
