% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_cluster.R
\name{.run_umap_prefilter}
\alias{.run_umap_prefilter}
\title{Run UMAP pre-cluster assignment script on host or in container}
\usage{
.run_umap_prefilter(fasta_fn, out_dir, min_samples = 50, sif_path = NULL)
}
\description{
Run UMAP pre-cluster assignment script on host or in container
}
\keyword{internal}
