% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_cluster.R
\name{.run_vsearch_cluster}
\alias{.run_vsearch_cluster}
\title{Run vsearch cluster_fast (MeShClust stand-in when meshclust binary unavailable)}
\usage{
.run_vsearch_cluster(fasta_fn, out_dir, id = 0.9, sif_path = NULL)
}
\description{
Run vsearch cluster_fast (MeShClust stand-in when meshclust binary unavailable)
}
\keyword{internal}
