% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_demultiplex.R
\name{.score_read_samples}
\alias{.score_read_samples}
\title{Score all samples for one read}
\usage{
.score_read_samples(read_seq, samples, primer_f, primer_r, end_window)
}
\description{
Score all samples for one read
}
\keyword{internal}
