% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refless_report.R
\name{.write_refless_html_report}
\alias{.write_refless_html_report}
\title{Write simple HTML quality report}
\usage{
.write_refless_html_report(out_dir, run_summary, sample_qc, demult_summary)
}
\description{
Write simple HTML quality report
}
\keyword{internal}
