% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basecall.R
\name{makeMMI}
\alias{makeMMI}
\title{Create minimap2 index for Dorado reference-guided basecalling}
\usage{
makeMMI(genome_fn, mmi_fn, minimap2_path)
}
\arguments{
\item{genome_fn}{Path to the reference genome FASTA file.}

\item{mmi_fn}{Output path for the minimap2 index (.mmi).}

\item{minimap2_path}{Path to the minimap2 executable.}
}
\description{
Create minimap2 index for Dorado reference-guided basecalling
}
