% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_amplicon_db.R
\name{prepAmpliconDB}
\alias{prepAmpliconDB}
\title{Prepare amplicon database for read alignment}
\usage{
prepAmpliconDB(blast_path, primer_list, genome_fn, out_dir, n_core)
}
\arguments{
\item{blast_path}{Path to a directory containing BLAST executables (makeblastdb and blastn).}

\item{primer_list}{Path to a CSV file containing primer sequences.
The file should have two columns: primer ID and primer sequence.
Forward primers should have IDs ending with "_F" and reverse primers
with IDs ending with "_R".}

\item{genome_fn}{Path to the reference genome sequence in FASTA format.}

\item{out_dir}{Output directory where the amplicon database will be created.}

\item{n_core}{Number of CPU cores to use for BLAST.}
}
\value{
Path to the generated amplicon FASTA file.
}
\description{
This function prepares an amplicon database for read alignment
by extracting amplicon sequences from a reference genome based
on provided primer sequences. It uses BLAST to identify the
locations of the primers in the reference genome and
extracts the corresponding amplicon sequences.
}
\examples{
# blast_path <- "/path/to/blast/bin" # Path to a directory containing blast executables (makeblastdb and blastn)
# primer_list <- "inst/extdata/amplicon_primers.csv"
# genome_fn <- '/reference/genome/sequence.fa' # Reference genome sequence in fasta format
# out_dir <- "/path/to/your/working/directory/output_directory_name"
# n_core <- 30 # Number of CPU cores to use
# amplicon_fn <- prepAmpliconDB(blast_path = blast_path,
#                               primer_list = primer_list,
#                               genome_fn = genome_fn,
#                               out_dir = out_dir,
#                               n_core = n_core)

}
