% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doAlign}
\alias{doAlign}
\title{Align reads to amplicon sequences using BLAST
This function aligns basecalled reads to a provided amplicon database using BLAST.
It processes the basecalled reads, performs the alignment, and extracts relevant alignment information.}
\usage{
doAlign(
  blast_path,
  basecall_fn,
  demult_out,
  amplicon_fn,
  align_dir,
  primer_list,
  pam_list,
  genome_fn,
  check_window = 10
)
}
\arguments{
\item{blast_path}{Path to a directory containing BLAST executables (makeblastdb and blastn).}

\item{basecall_fn}{A character vector containing paths to the FASTA files of basecalled reads.}

\item{demult_out}{A data frame summarizing the demultiplexing results.}

\item{amplicon_fn}{Path to the amplicon database FASTA file.}

\item{align_dir}{Path to the output directory where alignment results will be saved.}

\item{primer_list}{Path to a CSV file containing primer sequences.}

\item{pam_list}{Path to a CSV file containing PAM site information.}

\item{genome_fn}{Path to the reference genome sequence in FASTA format.}

\item{check_window}{An integer specifying the window size (in bp) around the expected cut site to search for edits.}
}
\value{
A data frame summarizing the alignment results.
}
\description{
Align reads to amplicon sequences using BLAST
This function aligns basecalled reads to a provided amplicon database using BLAST.
It processes the basecalled reads, performs the alignment, and extracts relevant alignment information.
}
