% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doBasecall}
\alias{doBasecall}
\title{Basecalling using Dorado}
\usage{
doBasecall(
  in_dir,
  basecall_dir,
  size_sel = c(0, Inf),
  dorado_path,
  samtools_path,
  n_core
)
}
\arguments{
\item{in_dir}{Path to the input directory containing raw sequencing data in pod5 format.}

\item{basecall_dir}{Path to the output directory where basecalling results will be saved.}

\item{size_sel}{A numeric vector of length 2 specifying the minimum and maximum read lengths (in bp) to retain after size selection.}

\item{dorado_path}{Path to the Dorado executable for basecalling.}

\item{samtools_path}{Path to the Samtools executable for BAM file processing.}

\item{n_core}{Number of CPU cores to use for parallel processing.}
}
\value{
A character vector containing paths to the size-selected FASTA files of basecalled reads.
}
\description{
This function performs basecalling on raw sequencing data using the Dorado basecaller.
It processes the raw data in a specified input directory, applies quality filtering,
and size selection, and outputs the basecalled reads in FASTA format.
}
