% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doDemultiplex}
\alias{doDemultiplex}
\title{Demultiplexing using BLAST}
\usage{
doDemultiplex(blast_path, basecall_fn, demult_dir, index_list)
}
\arguments{
\item{blast_path}{Path to a directory containing BLAST executables (makeblastdb and blastn).}

\item{basecall_fn}{A character vector containing paths to the FASTA files of basecalled reads.}

\item{demult_dir}{Path to the output directory where demultiplexing results will be saved.}

\item{index_list}{Path to a CSV file containing index sequences for demultiplexing.
The file should have five columns: index ID, forward index ID, forward index sequence, reverse index ID, and reverse index sequence.}
}
\value{
A data frame summarizing the demultiplexing results.
}
\description{
This function performs demultiplexing of basecalled reads using BLAST.
It matches the reads against provided index sequences to assign them to their respective samples.
}
