% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{evalMiao}
\alias{evalMiao}
\title{Evaluate MIAO results}
\usage{
evalMiao(out_dir, output_reads)
}
\arguments{
\item{out_dir}{The output directory where MIAO results are stored.}

\item{output_reads}{Logical. If TRUE, outputs the sequences of non-indexed aligned reads to a FASTA file.}
}
\value{
A summary report saved in the output directory.
}
\description{
This function evaluates the results of the MIAO pipeline and generates a summary report.
}
